firebase with home

+++ app/src/main/java/com/example/loboandchoco/Feed.kt
@@ -1,4 +1,4 @@
 package com.example.loboandchoco
 
-class Feed(val userId: String, val imageUrl: String, val profileImageUrl: String, var likeCount: Int, var isLIke: Boolean, var isBookmark: Boolean) {
+class Feed(val userId: String, val imageUrl: String, val profileImageUrl: String, var likeCount: Long, var isLIke: Boolean, var isBookmark: Boolean) {
 }
